Update gemfile, move twilio helper to app controller, update plans controller.
