+++ app/controllers/application_controller.rb
@@ -2,7 +2,7 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  helper_method :log_in, :current_user, :logged_in?, :log_out
+  helper_method :current_user, :logged_in?
 
   # Authentication methods
   def log_in(user)
@@ -24,5 +24,19 @@ class ApplicationController < ActionController::Base
   def log_out
     session.delete(:user_id)
     @current_user = nil
+  end
+
+  def send_twilio_notification(recipient, sender, message_body)
+    account_sid = ENV['TWILIO_SID']
+    auth_token = ENV['TWILIO_TOKEN']
+
+    # set up a client to talk to the Twilio REST API
+    @client = Twilio::REST::Client.new account_sid, auth_token
+
+    @client.account.messages.create({
+      :to => recipient,
+      :from => sender,
+      :body => message_body
+    })
   end
 end
