+++ app/controllers/plans_controller.rb
@@ -1,4 +1,5 @@
 class PlansController < ApplicationController
+
   def index
     @user = current_user
     @plans = @user.plans
@@ -40,6 +41,10 @@ class PlansController < ApplicationController
     # puts '*************************'
     # p @data
     # puts '*************************'
+
+    @message_body = "Greetings from Team Codex, #{@user.username}! Your new plan \'#{name}\' has been created. Login to check it out!"
+
+    send_twilio_notification("+12026572604", "+12027190379", @message_body)
 
     redirect_to action: "show", id: new_plan.id
   end
