+++ app/helpers/sessions_helper.rb
@@ -1,5 +1,2 @@
 module SessionsHelper
-
-
-
 end
